refactor: P1-T2 - 移动AVP入口源码到允许目录
