+++ LonelyPianistAVP/Resources/Info.plist
